Move DiagramItem to UML package
It shows a stereotype.
