+++ gaphor/SysML/diagramtype.py
@@ -1,9 +1,9 @@
 from gaphor.diagram.diagramtoolbox import DiagramType
-from gaphor.diagram.general.diagramitem import DiagramItem
 from gaphor.diagram.group import change_owner
 from gaphor.diagram.support import represents
 from gaphor.i18n import gettext
 from gaphor.SysML.sysml import SysMLDiagram
+from gaphor.UML.general import DiagramItem
 
 represents(SysMLDiagram)(DiagramItem)
