+++ gaphor/UML/general/__init__.py
@@ -1,4 +1,7 @@
 from gaphor.UML.general import connectors as connectors
 from gaphor.UML.general.comment import CommentItem as CommentItem
 from gaphor.UML.general.comment import CommentLineItem as CommentLineItem
+from gaphor.UML.general.diagramitem import DiagramItem as DiagramItem
 from gaphor.UML.general.image import ImageItem as ImageItem
+
+__all__ = ["CommentItem", "CommentLineItem", "DiagramItem", "ImageItem"]
