+++ gaphor/diagram/general/diagramitems.py
@@ -1,11 +1,9 @@
 # ruff: noqa: F401
 
-from gaphor.diagram.general.diagramitem import DiagramItem
 from gaphor.diagram.general.metadata import MetadataItem
 from gaphor.diagram.general.simpleitem import Box, Ellipse, Line
 
 __all__ = [
-    "DiagramItem",
     "Box",
     "Ellipse",
     "Line",
