+++ gaphor/diagram/tests/test_copypaste_full.py
@@ -1,9 +1,9 @@
 from gaphor import UML
 from gaphor.core.modeling import Diagram, ElementFactory
 from gaphor.diagram.copypaste import copy_full, paste_full
-from gaphor.diagram.general import DiagramItem
 from gaphor.diagram.tests.test_copypaste_link import two_classes_and_a_generalization
 from gaphor.UML.classes import ClassItem, GeneralizationItem, PackageItem
+from gaphor.UML.general import DiagramItem
 
 
 def test_copied_item_references_new_model_element(diagram, element_factory):
